lab3_middleware/project-dynamic-ice/pyproject.toml
[project]
name = "project-dynamic-ice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "zeroc-ice>=3.7.10.1",
]
